+++ src/i_colors.h
@@ -47,6 +47,7 @@
 extern byte     nearestcolors[PALETTESIZE];
 
 void I_InitTintTables(byte *palette);
+int FindNearestColor(byte *palette, int red, int green, int blue);
 void FindNearestColors(byte *palette);
 
 int FindDominantColor(patch_t *patch);